+++ packages/interact-with-bigquery/lib/bigquery_dialect.ts
@@ -1,14 +1,19 @@
 import {Dialect} from '@fairscript/interact/lib/databases/dialects'
 import {SelectStatement} from '@fairscript/interact/lib/statements/select_statement'
 import {GroupSelectStatement} from '@fairscript/interact/lib/statements/group_select_statement'
-import {adaptDistinct} from '@fairscript/interact/dist/lib/databases/ensuring_columns_in_select_distinct_clause_are_referenced_in_order_clause'
-import {Value} from '@fairscript/interact/lib/value'
-import {ValueRecord} from '@fairscript/interact/lib/record'
+import {ensureColumnsInSelectDistinctClauseAreReferencedInOrderClauseRule} from '@fairscript/interact/lib/databases/ensuring_columns_in_select_distinct_clause_are_referenced_in_order_clause'
+
+const statementAdaptationRules: ((statement: (SelectStatement|GroupSelectStatement)) => (SelectStatement|GroupSelectStatement))[] = [
+    ensureColumnsInSelectDistinctClauseAreReferencedInOrderClauseRule
+]
 
 export const bigQueryDialect: Dialect = {
     aliasEscape: '`',
     namedParameterPrefix: '@',
     useNamedParameterPrefixInRecord: false,
+
+    clientBooleanResultType: 'boolean',
+    clientFloatResultType: 'number',
 
     generateConvertToInteger(getColumn: string): string {
         return `CAST(${getColumn} AS INT64)`;
@@ -19,29 +24,9 @@ export const bigQueryDialect: Dialect = {
     },
 
     adaptSelectStatement(statement: SelectStatement|GroupSelectStatement): SelectStatement|GroupSelectStatement {
-        if(statement.kind === 'select-statement' && statement.distinct) {
-            return adaptDistinct(statement)
-        }
-        else {
-            return statement
-        }
-    },
-    adaptRows<T>(promisedResult: Promise<T[]>, adaptedSelectStatement: SelectStatement | GroupSelectStatement): Promise<T[]> {
-        return promisedResult
-    },
-    adaptScalar<T extends Value>(promisedResult: Promise<T>, adaptedSelectStatement: SelectStatement | GroupSelectStatement): Promise<T> {
-        return promisedResult
-    },
-    adaptSetOfRows<T extends { [p: string]: ValueRecord }>(promisedResult: Promise<T>, adaptedSelectStatement: SelectStatement | GroupSelectStatement): Promise<T> {
-        return promisedResult
-    },
-    adaptSetsOfRows<T extends { [p: string]: ValueRecord }>(promisedResult: Promise<T[]>, adaptedSelectStatement: SelectStatement | GroupSelectStatement): Promise<T[]> {
-        return promisedResult
-    },
-    adaptSingleRow<T extends ValueRecord>(promisedResult: Promise<T>, adaptedSelectStatement: SelectStatement | GroupSelectStatement): Promise<T> {
-        return promisedResult
-    },
-    adaptVector<T extends Value>(promisedResult: Promise<T[]>, adaptedSelectStatement: SelectStatement | GroupSelectStatement): Promise<T[]> {
-        return promisedResult
+        return statementAdaptationRules.reduce(
+            (adaptedStatement, rule) => rule(adaptedStatement),
+            statement
+        )
     }
 }